ship arena seperated from painterly
