+++ src/render/pipelines/std-painterly.ts
